Move Reserve Transactions flat file.
