+++ src/Libraries/Domains/Vendors/Confie.WesternGeneral/Confie.WesternGeneral.csproj
@@ -57,7 +57,6 @@
     <Compile Include="Feature.cs" />
     <Compile Include="PaymentTransaction.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="FlatFile\ReserveTransaction.cs" />
     <Compile Include="ReserveTransaction.cs" />
   </ItemGroup>
   <ItemGroup>
