+++ src/Libraries/Domains/Vendors/Confie.WesternGeneral/ReserveTransaction.cs
@@ -1,20 +1,60 @@
 ﻿using System;
+using FileHelpers;
+using FileHelpers.Events;
 
 namespace Confie.WesternGeneral
 {
-    public class ReserveTransaction
+    [FixedLengthRecord]
+    public class ReserveTransaction : INotifyRead
     {
-        public int ReserveTransactionId { get; set; }
+        [FieldFixedLength(50)]
+        [FieldTrim(TrimMode.Right)]
+        public string ClaimId { get; set; }
+
+        [FieldFixedLength(50)]
+        [FieldTrim(TrimMode.Right)]
+        public string FeatureId { get; set; }
+
+        [FieldFixedLength(18)]
+        [FieldConverter(ConverterKind.Date, "yyyy/MM/dd hh:mmtt")]
         public DateTime ReserveChangeDate { get; set; }
+
+        [FieldFixedLength(17)]
+        [FieldConverter(ConverterKind.Decimal)]
         public decimal ReserveBefore { get; set; }
+
+        [FieldFixedLength(17)]
+        [FieldConverter(ConverterKind.Decimal)]
         public decimal ReserveAfter { get; set; }
+
+        [FieldFixedLength(20)]
+        [FieldOptional]
+        [FieldTrim(TrimMode.Both)]
+        public string Dummy { get; set; }
+
+        [FieldHidden]
+        public int ReserveTransactionId { get; set; }
+
+        [FieldHidden]
         public string UpdatedUser { get; set; }
+
+        [FieldHidden]
         public string UpdatedDate { get; set; }
 
-        public string ClaimId { get; set; }
+        [FieldHidden]
         public virtual Claim Claim { get; set; }
 
-        public string FeatureId { get; set; }
+        [FieldHidden]
         public virtual Feature Feature { get; set; }
+
+        public void BeforeRead(BeforeReadEventArgs e)
+        {
+            e.SkipThisRecord = string.IsNullOrWhiteSpace(e.RecordLine) || e.RecordLine.Equals("\u001a");
+        }
+
+        public void AfterRead(AfterReadEventArgs e)
+        {
+            //Do nothing.
+        }
     }
 }
